requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
chromedriver-binary==96.0.4664.35.0
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
environ==1.0
et-xmlfile==1.1.0
h11==0.12.0
idna==3.3
numpy==1.21.4
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.4
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.2
wget==3.2
wsproto==1.0.0
